[School] SchoolMeal - 급식 먹는 로직 API 일부 구현

See merge request s11-final/S11P31E103!42

+++ Ssook_BE/src/main/java/com/stillalive/Ssook_BE/diner/service/DinerService.java
@@ -40,8 +40,8 @@ public class DinerService {
 
 
     public DinerListResDto getNearbyDiners(NearbyDinerResDto nearbyDinerResDto) {
-        Float lat = nearbyDinerResDto.getLat();
-        Float lng = nearbyDinerResDto.getLng();
+        Double lat = nearbyDinerResDto.getLat();
+        Double lng = nearbyDinerResDto.getLng();
         Float range = nearbyDinerResDto.getRange();
 
         return DinerListResDto.builder()
